NLP/src/requirements.txt
annotated-types        0.7.0
anthropic              0.34.2
anyio                  4.6.0
blinker                1.8.2
certifi                2024.8.30
charset-normalizer     3.3.2
click                  8.1.7
colorama               0.4.6
distro                 1.9.0
filelock               3.13.1
Flask                  3.0.3
fsspec                 2024.2.0
h11                    0.14.0
httpcore               1.0.6
httpx                  0.27.2
huggingface-hub        0.25.1
idna                   3.10
itsdangerous           2.2.0
Jinja2                 3.1.3
jiter                  0.5.0
joblib                 1.4.2
MarkupSafe             2.1.5
mpmath                 1.3.0
networkx               3.2.1
nltk                   3.9.1
numpy                  1.26.3
packaging              24.1
pillow                 10.2.0
pip                    24.2
py-readability-metrics 1.4.5
pydantic               2.9.2
pydantic_core          2.23.4
python-dateutil        2.9.0.post0
PyYAML                 6.0.2
regex                  2024.9.11
requests               2.32.3
safetensors            0.4.5
scikit-learn           1.5.2
scipy                  1.14.1
setuptools             70.0.0
six                    1.16.0
sniffio                1.3.1
sympy                  1.12
threadpoolctl          3.5.0
tokenizers             0.20.0
torch                  2.4.1+cu124
torchaudio             2.4.1+cu124
torchvision            0.19.1+cu124
tqdm                   4.66.5
transformers           4.45.1
typing_extensions      4.9.0
urllib3                2.2.3
Werkzeug               3.0.4
